Move type logic in schema_metadata.js to metabase-lib
