+++ frontend/src/metabase/lib/formatting/date.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import moment, { Moment } from "moment-timezone";
 
 import { parseTimestamp } from "metabase/lib/time";
-import { isDateWithoutTime } from "metabase/lib/schema_metadata";
 import type { DatetimeUnit } from "metabase-types/api/query";
+import { isDateWithoutTime } from "metabase-lib/lib/types/utils/isa";
 import {
   DEFAULT_DATE_STYLE,
   DEFAULT_TIME_STYLE,
