+++ frontend/src/metabase/lib/formatting/geography.ts
@@ -1,7 +1,7 @@
 import d3 from "d3";
 
-import { isLatitude, isLongitude } from "metabase/lib/schema_metadata";
 import { decimalCount } from "metabase/visualizations/lib/numeric";
+import { isLatitude, isLongitude } from "metabase-lib/lib/types/utils/isa";
 import { OptionsType } from "./types";
 
 const DECIMAL_DEGREES_FORMATTER = d3.format(".08f");
