+++ frontend/src/metabase/lib/formatting/link.ts
@@ -1,7 +1,6 @@
-import { isDate } from "metabase/lib/schema_metadata";
-
 import { formatValue } from "metabase/lib/formatting";
 import type { DatasetColumn } from "metabase-types/api/dataset";
+import { isDate } from "metabase-lib/lib/types/utils/isa";
 import { formatDateTimeForParameter } from "./date";
 
 interface TemplateForClickFormatFunctionParamsType {
