+++ frontend/src/metabase/query_builder/components/filters/modals/BulkFilterList/utils.ts
@@ -1,7 +1,5 @@
-import { isDate } from "metabase/lib/schema_metadata";
+import { isDate } from "metabase-lib/lib/types/utils/isa";
 import { DimensionOption } from "metabase-lib/lib/queries/StructuredQuery";
-
-import { LONG_TEXT_MIN } from "metabase-lib/lib/metadata/Field";
 
 type PriorityMap = { [key: string]: number | undefined };
