+++ frontend/src/metabase/query_builder/components/filters/pickers/DefaultPicker.jsx
@@ -9,11 +9,11 @@ import FieldValuesWidget from "metabase/components/FieldValuesWidget";
 import {
   getFilterArgumentFormatOptions,
   isFuzzyOperator,
-  isCurrency,
 } from "metabase/lib/schema_metadata";
 
 import { getCurrencySymbol } from "metabase/lib/formatting";
 
+import { isCurrency } from "metabase-lib/lib/types/utils/isa";
 import { keyForColumn } from "metabase-lib/lib/queries/utils/dataset";
 import TextPicker from "./TextPicker";
 import SelectPicker from "./SelectPicker";
