+++ frontend/src/metabase/visualizations/components/ObjectDetail/utils.ts
@@ -2,9 +2,9 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import { singularize } from "metabase/lib/formatting";
-import { isPK, isEntityName } from "metabase/lib/schema_metadata";
 
 import { DatasetData, Column } from "metabase-types/types/Dataset";
+import { isPK, isEntityName } from "metabase-lib/lib/types/utils/isa";
 import Question from "metabase-lib/lib/Question";
 import Table from "metabase-lib/lib/metadata/Table";
