+++ frontend/src/metabase/visualizations/lib/settings/utils.js
@@ -1,10 +1,10 @@
 import _ from "underscore";
-import { isDimension, isMetric } from "metabase/lib/schema_metadata";
 import {
   getFriendlyName,
   columnsAreValid,
   getDefaultDimensionAndMetric,
 } from "metabase/visualizations/lib/utils";
+import { isDimension, isMetric } from "metabase-lib/lib/types/utils/isa";
 
 export function getOptionFromColumn(col) {
   return {
