+++ frontend/src/metabase/visualizations/lib/timeseries.js
@@ -1,8 +1,8 @@
 import moment from "moment-timezone";
 import _ from "underscore";
 
-import { isDate } from "metabase/lib/schema_metadata";
 import { parseTimestamp } from "metabase/lib/time";
+import { isDate } from "metabase-lib/lib/types/utils/isa";
 
 import { unexpectedTimezoneWarning, multipleTimezoneWarning } from "./warnings";
