+++ frontend/src/metabase/visualizations/lib/utils.js
@@ -3,7 +3,11 @@ import d3 from "d3";
 import { t } from "ttag";
 import crossfilter from "crossfilter";
 
-import { isDimension, isMetric, isDate } from "metabase/lib/schema_metadata";
+import {
+  isDimension,
+  isMetric,
+  isDate,
+} from "metabase-lib/lib/types/utils/isa";
 
 export const MAX_SERIES = 100;
