+++ frontend/src/metabase/visualizations/visualizations/SmartScalar.jsx
@@ -3,7 +3,6 @@ import React from "react";
 import { t, jt } from "ttag";
 import _ from "underscore";
 
-import { isDate } from "metabase/lib/schema_metadata";
 import { formatNumber, formatValue } from "metabase/lib/formatting";
 import { color } from "metabase/lib/colors";
 
@@ -16,6 +15,7 @@ import ScalarValue, {
   ScalarWrapper,
   ScalarTitle,
 } from "metabase/visualizations/components/ScalarValue";
+import { isDate } from "metabase-lib/lib/types/utils/isa";
 import { formatBucketing } from "metabase-lib/lib/queries/utils/query-time";
 
 import {
